Update README.md
Conclusion:
In this paper, an optimized particle swarm optimization algorithm is used to calculate the switching angles of symmetric cascade multilevel inverters to satisfy optimal first order harmonics and reduce undesirable harmonics. Due to the need for high power inverters and the growing use of these types of transformers in the power system, the use of multilevel inverters is increasing. In this inverter, an increase in the number of dc power sources can be achieved at higher voltages, while an increase in the number of dc power sources can be achieved by properly adjusting the switching angles of the total harmonic distortion. When the number of switching angles increases, conventional methods such as Newton-Raphson are not capable of solving algebraic equations for computing angles; meta-heuristic algorithms can be used to solve the problem of determining angles of fire. Comparing the results of the PSO and IPSO methods, it was found that for the separate dc power supplies the results obtained from the IPSO method are far better than the PSO method. Also, the performance of IPSO algorithm in determining switching angles for five to nineteen level inverters was compared with PSO algorithm.